fix: spaces containing in in -o DIR was mixed with automatically added quotes
The output_dir variable is build from user input and the parallel
special sequences, which already contain quotes. Therefore double
quoting would fail.

So I quote the variable halfway through only partially and not quote
it again when actually using output_dir in the 7z command, as it still
contains the special sequences that will be replaced with quoted parts
at the last moment when command is run.